core, xeth: implemented auto resend strategy for owned transactions
Added an new transaction type used by the transaction pool `poolTx`.
This struct hold information about the transaction state:

* amount of post tries
* added timestamp
* last post timestamp

Whenever a transaction is being submitted by the user the transaction is
flagged as owned. When an owned transaction sits in the pool for too
long (39 seconds) it will be posted again.

A future PR could check the amount of post tries and will post back a
failure event. An failure event could potentially be posted back to a
frontend connected to the `geth` instance and deletes from the pool. A
frontend could ask the user for a higher gas price or some other action.